More equip-screen; more equipping logic.
The equipment screen is fairly functional: can select a piece of
equipment to equip, works with multiple accessories and two one- or one
two-handed weapon(s). There's no inventory yet so you have no pool to
choose from or unequip to, and there's no stats to improve with these
items, but we're getting there.